--- cookbook/transformers/fsdp2_moe.sh
@@ -1,1 +1,0 @@
-CUDA_VISIBLE_DEVICES=0,1,2,3,4,5,6,7 torchrun --nproc_per_node=8 fsdp2_moe.py
